requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.7.0
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
django-json-widget==2.0.1
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
